requirements.txt
# requirements.txt
pandas>=2.0.0
numpy>=1.24.0
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.3
hdbscan>=0.8.33
scikit-learn>=1.3.0
plotly>=5.14.0
wordcloud>=1.9.0
nltk>=3.8.0
Sastrawi>=1.0.1
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0